adding database and cache persistence.

+++ src/main/java/org/platinumrx/pgrouting/rto/TransactionInitiateRequest.java
@@ -3,13 +3,17 @@ package org.platinumrx.pgrouting.rto;
 import lombok.*;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Setter
 @ToString
 public class TransactionInitiateRequest {
-    @JsonProperty("order_id") private String orderId;
-    public Double amount;
-    @JsonProperty("payment_instrument") private PaymentInstrument paymentInstrument;
+    @JsonProperty("order_id")
+    private String orderId;
+    public BigDecimal amount;
+    @JsonProperty("payment_instrument")
+    private PaymentInstrument paymentInstrument;
 }
